Adjusted, restructured, and created routing files to move all authentication and comment routes to API folder with proper controllers used
